+++ data/client/client_test.go
@@ -14,6 +14,7 @@ import (
 	dataClient "github.com/tidepool-org/platform/data/client"
 	"github.com/tidepool-org/platform/id"
 	"github.com/tidepool-org/platform/platform"
+	testHTTP "github.com/tidepool-org/platform/test/http"
 )
 
 var _ = Describe("Client", func() {
@@ -23,7 +24,8 @@ var _ = Describe("Client", func() {
 		BeforeEach(func() {
 			config = platform.NewConfig()
 			Expect(config).ToNot(BeNil())
-			config.Address = "http://localhost:1234"
+			config.Address = testHTTP.NewAddress()
+			config.UserAgent = testHTTP.NewUserAgent()
 			config.Timeout = 30 * time.Second
 		})
 
@@ -40,6 +42,13 @@ var _ = Describe("Client", func() {
 			Expect(clnt).To(BeNil())
 		})
 
+		It("returns an error if config user agent is missing", func() {
+			config.UserAgent = ""
+			clnt, err := dataClient.New(config)
+			Expect(err).To(MatchError("config is invalid; user agent is missing"))
+			Expect(clnt).To(BeNil())
+		})
+
 		It("returns success", func() {
 			clnt, err := dataClient.New(config)
 			Expect(err).ToNot(HaveOccurred())
@@ -49,14 +58,17 @@ var _ = Describe("Client", func() {
 
 	Context("with started server and new client", func() {
 		var server *Server
+		var userAgent string
 		var clnt dataClient.Client
 		var ctx context.Context
 
 		BeforeEach(func() {
 			server = NewServer()
+			userAgent = testHTTP.NewUserAgent()
 			config := platform.NewConfig()
 			Expect(config).ToNot(BeNil())
 			config.Address = server.URL()
+			config.UserAgent = userAgent
 			config.Timeout = 30 * time.Second
 			var err error
 			clnt, err = dataClient.New(config)
@@ -101,6 +113,7 @@ var _ = Describe("Client", func() {
 						server.AppendHandlers(
 							CombineHandlers(
 								VerifyRequest("DELETE", fmt.Sprintf("/v1/users/%s/data", userID)),
+								VerifyHeaderKV("User-Agent", userAgent),
 								VerifyHeaderKV("X-Tidepool-Session-Token", token),
 								VerifyBody([]byte{}),
 								RespondWith(http.StatusUnauthorized, nil, nil)),
@@ -119,6 +132,7 @@ var _ = Describe("Client", func() {
 						server.AppendHandlers(
 							CombineHandlers(
 								VerifyRequest("DELETE", fmt.Sprintf("/v1/users/%s/data", userID)),
+								VerifyHeaderKV("User-Agent", userAgent),
 								VerifyHeaderKV("X-Tidepool-Session-Token", token),
 								VerifyBody([]byte{}),
 								RespondWith(http.StatusForbidden, nil, nil)),
@@ -137,6 +151,7 @@ var _ = Describe("Client", func() {
 						server.AppendHandlers(
 							CombineHandlers(
 								VerifyRequest("DELETE", fmt.Sprintf("/v1/users/%s/data", userID)),
+								VerifyHeaderKV("User-Agent", userAgent),
 								VerifyHeaderKV("X-Tidepool-Session-Token", token),
 								VerifyBody([]byte{}),
 								RespondWith(http.StatusOK, nil, nil)),