+++ notification/client/client_test.go
@@ -8,6 +8,7 @@ import (
 	"github.com/tidepool-org/platform/notification"
 	notificationClient "github.com/tidepool-org/platform/notification/client"
 	"github.com/tidepool-org/platform/platform"
+	testHTTP "github.com/tidepool-org/platform/test/http"
 )
 
 var _ = Describe("Client", func() {
@@ -20,7 +21,8 @@ var _ = Describe("Client", func() {
 
 	Context("New", func() {
 		BeforeEach(func() {
-			cfg.Address = "http://localhost:1234"
+			cfg.Address = testHTTP.NewAddress()
+			cfg.UserAgent = testHTTP.NewUserAgent()
 		})
 
 		It("returns an error if unsuccessful", func() {
@@ -38,12 +40,15 @@ var _ = Describe("Client", func() {
 
 	Context("with server and new client", func() {
 		var svr *Server
+		var userAgent string
 		var clnt notification.Client
 
 		BeforeEach(func() {
 			svr = NewServer()
+			userAgent = testHTTP.NewUserAgent()
 			Expect(svr).ToNot(BeNil())
 			cfg.Address = svr.URL()
+			cfg.UserAgent = userAgent
 			var err error
 			clnt, err = notificationClient.New(cfg)
 			Expect(err).ToNot(HaveOccurred())