+++ service/service/DEPRECATED_service.go
@@ -69,6 +69,8 @@ func (d *DEPRECATEDService) initializeAuthClient() error {
 	d.Logger().Debug("Loading auth client config")
 
 	cfg := authClient.NewConfig()
+	cfg.UserAgent = d.UserAgent()
+	cfg.ExternalConfig.UserAgent = d.UserAgent()
 	if err := cfg.Load(d.ConfigReporter().WithScopes("auth", "client")); err != nil {
 		return errors.Wrap(err, "unable to load auth client config")
 	}