+++ user/service/service/standard.go
@@ -131,6 +131,7 @@ func (s *Standard) initializeDataClient() error {
 	s.Logger().Debug("Loading data client config")
 
 	cfg := platform.NewConfig()
+	cfg.UserAgent = s.UserAgent()
 	if err := cfg.Load(s.ConfigReporter().WithScopes("data", "client")); err != nil {
 		return errors.Wrap(err, "unable to load data client config")
 	}
@@ -150,6 +151,7 @@ func (s *Standard) initializeMetricClient() error {
 	s.Logger().Debug("Loading metric client config")
 
 	cfg := platform.NewConfig()
+	cfg.UserAgent = s.UserAgent()
 	if err := cfg.Load(s.ConfigReporter().WithScopes("metric", "client")); err != nil {
 		return errors.Wrap(err, "unable to load metric client config")
 	}
@@ -169,6 +171,7 @@ func (s *Standard) initializeUserClient() error {
 	s.Logger().Debug("Loading user client config")
 
 	cfg := platform.NewConfig()
+	cfg.UserAgent = s.UserAgent()
 	if err := cfg.Load(s.ConfigReporter().WithScopes("user", "client")); err != nil {
 		return errors.Wrap(err, "unable to load user client config")
 	}